+++ src/options.rs
@@ -24,6 +24,9 @@ pub struct Options {
     /// Whether to respect `.fdignore` files or not.
     pub read_fdignore: bool,
 
+    /// Whether to respect ignore files in parent directories or not.
+    pub read_parent_ignore: bool,
+
     /// Whether to respect VCS ignore files (`.gitignore`, ..) or not.
     pub read_vcsignore: bool,
 
@@ -53,6 +56,10 @@ pub struct Options {
 
     /// The number of threads to use.
     pub threads: usize,
+
+    /// If true, the program doesn't print anything and will instead return an exit code of 0
+    /// if there's at least one match. Otherwise, the exit code will be 1.
+    pub quiet: bool,
 
     /// Time to buffer results internally before streaming to the console. This is useful to
     /// provide a sorted output, in case the total execution time is shorter than